Added automatic intention detection.
Intentions get added to the plot graph whenever an agent puts a plan
onto its intention stack. (see PlotAwareAg#selectOption)
The vertex of speech acts was changed to have the format of an
intention. (see PlotGraphController#addMsgSend)
The vertex of listen events was changed to have the format of a percept.
(see PlotGraphController#addMsgReceive and Vertex#toString)