Added indentation for motivation edges and fixed adding edges
Fixed a bug where motivation/actualization edges would not show up in
the graph because of a duplicate temporal edge by extending
PlotDirectedSparseGraph from DirectedSparseMultigraph.
Added indentation to motivation edges in order to increase readability
of the graph.
Removed yellow coloring of vertices with motivation.
Changed color of actualization edges to a darker cyan.

Issues:
- Indentation offset is currently saved in the Edge class, but it should
rather be in a class corresponding to rendering of the graph.
- Actualization edge position is fixed to an arbitrary value (-20f).
This should be more dynamic to account for the possibility of small
vertices.