Fixed 2 appraisals in ASL code and added source-of-mood motivation
Fixed appraisal of disappointment emotion when failing to eat bread.
Added appraisal of reproach emotion when receiving help request but not
wanting to help.
Implemented using the source of the mood as motivation where motivation
would previously be plotted as "mood".

Known Issue: if the set of sources of a mood contains more than one
source, post-processing will fail. (currently not happening)