Implemented logging to std out in Pause Mode
Pause Mode in jason works by MASConsoleGUI checking on each append,
whether pause is on, and in that case executing wait(), until pause mode
is finished, which wakes the GUI up again. This means, environment and
agents continue running, until something wants to log (usually:
Environment#executeAction) which has to wait until the GUI thread wakes
up. This is clearly not the best idea, and crucially prevents logging in
while the simulation is pause.

Fixed this by making PlotEnvironment check, if pause mode is on, and
waiting if it is.