Added an equivalence functionality for vertices with opposing emotions
+ Added a second post-processing step (CompactGraphPPVisitor)
* Renamed first post-processing step (FullGraphPPVisitor)
+ A vertex with several emotions of opposing P gets split into two
vertices, one with all the negative emotions and one with all the
positive ones. This removes the need for equivalence edges per se.

Known issue:
	A vertex that is part of a functional unit and gets split creates the
same functional unit with polyvalence in all vertices but the one that
was split. This GREATLY affects tellability. Possibly not clone the
edges but create an equivalence edge after all?