Added honored request unit and fixed bugs
Added a unit for "honored request" based on the implementation of
"denied request" (simply swapped the negative to positive).
	This unit is actually found in stories where a character accepts the
request of the hen and should increase tellability in stories where two
or more characters accept the request.

Plot analysis bug fix:
	Polyvalent vertices now get marked with an asterisk after all
functional units were identified so as to not interfere with the results
	(Previously, intentions marked as polyvalent would not count as
intentions for the following analysis because of the beginning asterisk)

Edge layout bug fix:
	In some cases, vertices would not get resized according to all edges
associated with that vertex.
	Now, all edges are considered when setting a vertex' minimum width to
be used in the transformer.