Improved connectivity graph and framing
Note: translation of "help_with" to natural language does intentionally
not process the arguments to the term but rather stops (see Translator
class). This was done to obfuscate that only one set of denied request
units contributes to the summary, but can of course be changed easily at
any time by changing the "help"s to "help with"s and changing Form.STOP
to Form.ING.


Known issues:
	- In some configurations (i.e. with an honored request), Plotmas
crashes silently after pressing "Analyze graph". Presumably an exception
when generating the summary, which is not displayed in any console?
	- Plural/singular is hard-coded in the natural language snippets of the
unit. I.e. if only a single agent would deny the hen's request, the
output would be "the cow deny the hen's request" instead of "denies".