Employing CSP solver to identify AdaptPersonality fixes
+ using choco-solver
+ each personality-annotation is now parsed into CSP problem and soled
in the domain {-1, -0.3, 0.3,1}
+ for each plan that could get activated by intention, which is
triggered by unresolved happening, the different viable personality
settings are computed for each plan step. The disjunction of these
settings (since each one might allowthe step to be executed) is
conjuncted with each other (so that for each step a valid setting is
selected). This CSP is solved, and each solution is a viable
AdaptPersonality fix.

-JavaDoc not done properly
-Testing not done or out of date