Current state:
CounterfactualityLauncher works specific for RedHenLauncher. It calls
RedHenCounterfactualityCycle, where the best personalities are chosen.

Problem: At the moment I am working only with best tellability (evth
copy and paste) but something in the cycle is not working.

As soon as this is fixed, the following will be done:
Producing a plotdirectedsparse graph from the best personalities
Produce an own cycle for best counterfactuality