Fixed bug with inexact matching
State would mess up candidateV2 index when following predecessor edges,
because after a transformation it is necessary to increase candidateV2
index by 1 since another vertex was added before it during
transformation.

This was not necessary with successor edges, because in that case we
want candidateV2 index to point to the newly inserted vertex.