Refactored plot graph analysis to allow asynchronous execution
+ Implemented GraphAnalyzer (extending Thread), that initiates graph
post processing and tellability computation. It supports an asynchronous
mode as an own thread, and a synchronous mode for ER cycles and testing
purposes.
+ Introduced interface AnalysisResultListener, that has to be
implemented by every class that wants to start and receice grapha
analysis results (here: PlotControlsLauncher)
+ Refactored PlotGraphController so that it only displays results of
analysis, instead of initiating it

Bug Fix:
+ Refactored PlotGraphController so that displayed graphs are now
selected using name and not graph equality, since latter is slower and
indeterministically breaks on larger graphs